Added events on the homescreen
Consider the case that the API is going to change as we don't want to
expose the attendees and other data by default. We also need to handle
errors on the home screen properly.

In addition to this, the registration screen also needs to be more
robust in the case of errors and the home screen.