Add support for refresh tokens
The authentication mechanism has been changed significantly, from now on
we will need to maintain a refresh and an access token, requests if they
are unauthenticated will now check to refresh tokens.

There is one issue where every screen making a request now needs to deal
with having to redirect to the login screen if a call has an error. This
makes code harder to write and repeated and needs to be fixed soon.