Implement user details caching
A users role can always be changed so whenever the token which stores
the role updates we need to extract the latest role and other profile
preferences. This might not be the most efficient way as we need to make
a call to the users endpoint each time we refresh a token. We should
consider alternative strategies.