Make profile screen generic
The profile screen is now generic and will vary depending on whether
it's your own profile or not. If it is your own then you should be able
to update it. In addition to this, the parameters which can actually be
updated are fixed now. To change your email and password this will now
require separate endpoints and screens which aren't yet implemented.